Bug fixes and cleanup for integration of new Prearchive API
Former-commit-id: 2ffd983d975ec34ef8f294be9c1bd40d4b7aac3d

+++ plugin-resources/webapp/xnat/java/org/nrg/xnat/restlet/resources/prearchive/PrearcSessionResource.java
@@ -86,7 +86,7 @@ public final class PrearcSessionResource extends SecureResource {
 		if(project.equalsIgnoreCase("Unassigned"))project=null;
 		final File sessionDir;
 		try {
-			sessionDir = PrearcUtils.getPrearcSessionDir(user, project, timestamp, session);
+			sessionDir = PrearcUtils.getPrearcSessionDir(user, project, timestamp, session,false);
 		} catch (InvalidPermissionException e) {
 			logger.error("",e);
 			this.getResponse().setStatus(Status.CLIENT_ERROR_FORBIDDEN, e.getMessage());
@@ -101,7 +101,7 @@ public final class PrearcSessionResource extends SecureResource {
 		if (POST_ACTION_BUILD.equals(action)) {
 			try {
 				PrearcDatabase.buildSession(sessionDir, session, timestamp, project);
-				PrearcUtils.resetStatus(user, project, timestamp, session);
+				PrearcUtils.resetStatus(user, project, timestamp, session,true);
 			} catch (InvalidPermissionException e) {
 				logger.error("",e);
 				this.getResponse().setStatus(Status.CLIENT_ERROR_FORBIDDEN, e.getMessage());
@@ -112,7 +112,7 @@ public final class PrearcSessionResource extends SecureResource {
 		} else if (POST_ACTION_RESET.equals(action)) {
 			try {
 				PrearcDatabase.buildSession(sessionDir, session, timestamp, project);
-				PrearcUtils.resetStatus(user, project, timestamp, session);
+				PrearcUtils.resetStatus(user, project, timestamp, session,true);
 			} catch (InvalidPermissionException e) {
 				logger.error("",e);
 				this.getResponse().setStatus(Status.CLIENT_ERROR_FORBIDDEN, e.getMessage());
@@ -126,10 +126,12 @@ public final class PrearcSessionResource extends SecureResource {
 			if(StringUtils.isNotEmpty(newProj)){
 				//TODO: convert ALIAS to project ID (if necessary)
 			}
-			
-			//TODO: check permissions on new Project
-			
+
 			try {
+				if(!PrearcUtils.canModify(user, newProj)){
+					this.getResponse().setStatus(Status.CLIENT_ERROR_FORBIDDEN, "Unable to modify session data for destination project.");
+					return;
+				}
 				if(PrearcDatabase.setStatus(session, timestamp, project, PrearcStatus.MOVING)){
 					PrearcDatabase.moveToProject(session, timestamp, project, newProj);
 				}				
@@ -142,6 +144,9 @@ public final class PrearcSessionResource extends SecureResource {
 			} catch (SessionException e) {
 				logger.error("",e);
 				this.getResponse().setStatus(Status.CLIENT_ERROR_BAD_REQUEST, e);
+			} catch (Exception e) {
+				logger.error("",e);
+				this.getResponse().setStatus(Status.SERVER_ERROR_INTERNAL, e);
 			}			
 		} else {
 			this.getResponse().setStatus(Status.CLIENT_ERROR_BAD_REQUEST,
@@ -151,10 +156,12 @@ public final class PrearcSessionResource extends SecureResource {
 
 	@Override
 	public void handleDelete() {
+
+		if(project.equalsIgnoreCase("Unassigned"))project=null;
 		
 		try {
 			//checks if the user can access this session
-			PrearcUtils.getPrearcSessionDir(user, project, timestamp, session);
+			PrearcUtils.getPrearcSessionDir(user, project, timestamp, session,false);
 		} catch (InvalidPermissionException e) {
 			logger.error("",e);
 			this.getResponse().setStatus(Status.CLIENT_ERROR_FORBIDDEN, e.getMessage());
@@ -166,7 +173,18 @@ public final class PrearcSessionResource extends SecureResource {
 		}
 		
 		try {
-			PrearcDatabase.deleteSession(session, timestamp, project);
+			if(!PrearcUtils.canModify(user, project)){
+				this.getResponse().setStatus(Status.CLIENT_ERROR_FORBIDDEN, "Unable to modify session data for destination project.");
+				return;
+			}
+			
+			if(PrearcDatabase.setStatus(session, timestamp, project, PrearcStatus.DELETING)){
+				PrearcDatabase.deleteSession(session, timestamp, project);
+			}
+		} catch (SessionException e) {
+			logger.error("",e);
+			this.getResponse().setStatus(Status.CLIENT_ERROR_NOT_FOUND, e.getMessage());
+			return;
 		} catch (Exception e) {
 			logger.error("",e);
 			this.getResponse().setStatus(Status.SERVER_ERROR_INTERNAL, e.getMessage());
@@ -184,7 +202,7 @@ public final class PrearcSessionResource extends SecureResource {
 	public Representation getRepresentation(final Variant variant){
 		final File sessionDir;
 		try {
-			sessionDir = PrearcUtils.getPrearcSessionDir(user, project, timestamp, session);
+			sessionDir = PrearcUtils.getPrearcSessionDir(user, project, timestamp, session,false);
 		} catch (InvalidPermissionException e) {
 			this.getResponse().setStatus(Status.CLIENT_ERROR_FORBIDDEN, e.getMessage());
 			return null;
